requirements.txt
Flask>=2.0
python-dotenv
